refactor: move SidebarRegion type from types/tab.ts to types/layout.ts
SidebarRegion describes the app shell's four fixed regions and belongs
to the layout system, not the tab system. Create types/layout.ts and
update all imports accordingly.

Closes #238
