+++ spa/src/components/RegionContextMenu.tsx
@@ -2,7 +2,7 @@ import { useEffect, useLayoutEffect, useRef } from 'react'
 import { CheckSquare, Square } from '@phosphor-icons/react'
 import { useLayoutStore } from '../stores/useLayoutStore'
 import { getAllViews } from '../lib/module-registry'
-import type { SidebarRegion } from '../types/tab'
+import type { SidebarRegion } from '../types/layout'
 
 interface Props {
   region: SidebarRegion
