+++ spa/src/components/SidebarRegion.tsx
@@ -5,7 +5,7 @@ import { getViewDefinition } from '../lib/module-registry'
 import { RegionResize } from './RegionResize'
 import { RegionManager } from './RegionManager'
 import { RegionContextMenu } from './RegionContextMenu'
-import type { SidebarRegion as SidebarRegionType } from '../types/tab'
+import type { SidebarRegion as SidebarRegionType } from '../types/layout'
 import { useWorkspaceStore } from '../features/workspace/store'
 import { useHostStore } from '../stores/useHostStore'
 import { useTabStore } from '../stores/useTabStore'
