+++ spa/src/types/layout.ts
@@ -1,0 +1,6 @@
+// === Sidebar Region (layout system) ===
+export type SidebarRegion =
+  | 'primary-sidebar'
+  | 'primary-panel'
+  | 'secondary-panel'
+  | 'secondary-sidebar'
